[Decomp][Player] Cohesion split: extract item-apply system to PlayerItemApply.cpp
Moves the contiguous item-equip side-effects block (stat mods, weapon-
dependent auras, on-equip / on-store / use-item spell hooks, ammo
bonuses) out of Player.cpp into a new file under src/game/Object/.

Player.cpp: 18,889 -> 17,700 lines (-1,189).
PlayerItemApply.cpp: 1,272 lines, 16 methods.

Methods moved:
- _ApplyItemMods, _ApplyItemBonuses
- _ApplyWeaponDependentAuraMods, _ApplyWeaponDependentAuraCritMod,
  _ApplyWeaponDependentAuraDamageMod
- ApplyItemEquipSpell, ApplyEquipSpell, UpdateEquipSpellsAtFormChange
- ApplyItemOnStoreSpell, DestroyItemWithOnStoreSpell
- _HandleDeadlyPoison, CastItemCombatSpell, CastItemUseSpell
- _RemoveAllItemMods, _ApplyAllItemMods, _ApplyAllLevelScaleItemMods
- _ApplyAmmoBonuses, CheckAmmoCompatibility

No header change; no signature change; no behaviour change.
